+++ server/neptune/workflows/internal/deploy/request/converter/repo.go
@@ -1,0 +1,27 @@
+package converter
+
+import (
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/deploy/request"
+	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/github"
+)
+
+func Repo(external request.Repo) github.Repo {
+	return github.Repo{
+		Name:  external.Name,
+		Owner: external.Owner,
+		URL:   external.URL,
+		Credentials: github.AppCredentials{
+			InstallationToken: external.Credentials.InstallationToken,
+		},
+		Ref: github.Ref{
+			Name: external.Ref.Name,
+			Type: external.Ref.Type,
+		},
+	}
+}
+
+func User(external request.User) github.User {
+	return github.User{
+		Username: external.Name,
+	}
+}
