+++ server/neptune/workflows/internal/deploy/request/converter/root.go
@@ -2,7 +2,6 @@ package converter
 
 import (
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/deploy/request"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/github"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/job"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/root"
 )
@@ -23,27 +22,6 @@ func Root(external request.Root) root.Root {
 		Trigger:   root.Trigger(external.Trigger),
 	}
 
-}
-
-func Repo(external request.Repo) github.Repo {
-	return github.Repo{
-		Name:  external.Name,
-		Owner: external.Owner,
-		URL:   external.URL,
-		Credentials: github.AppCredentials{
-			InstallationToken: external.Credentials.InstallationToken,
-		},
-		Ref: github.Ref{
-			Name: external.Ref.Name,
-			Type: external.Ref.Type,
-		},
-	}
-}
-
-func User(external request.User) github.User {
-	return github.User{
-		Username: external.Name,
-	}
 }
 
 func mode(mode request.PlanMode) *job.PlanMode {
